# ddtrace/tracer: add NewStartSpanConfig & WithStartSpanConfig API