--- a/ddtrace/ddtrace.go
+++ b/ddtrace/ddtrace.go
@@ -171,6 +171,15 @@ type StartSpanConfig struct {
 	Context context.Context
 }
 
+func NewStartSpanConfig(opts ...StartSpanOption) StartSpanConfig {
+	var cfg StartSpanConfig
+	for _, fn := range opts {
+		fn(&cfg)
+	}
+
+	return cfg
+}
+
 // Logger implementations are able to log given messages that the tracer or profiler might output.
 type Logger interface {
 	// Log prints the given message.