--- a/ddtrace/ddtrace.go
+++ b/ddtrace/ddtrace.go
@@ -171,6 +171,8 @@ type StartSpanConfig struct {
 	Context context.Context
 }
 
+// NewStartSpanConfig allows to build a base config struct. It accepts the same options as StartSpan.
+// It's useful to reduce the number of operations in any hot path and update it for request/operation specifics.
 func NewStartSpanConfig(opts ...StartSpanOption) StartSpanConfig {
 	var cfg StartSpanConfig
 	for _, fn := range opts {