### What does this PR do?

Adds `NewStartSpanConfig` and `WithStartSpanConfig` functions to allow to use config structs.

### Motivation

This API reduces the number of calls (in functional option form) in hot paths by giving the freedom to prepare a common span configuration in hot paths.

Microbenchmarks show that there are significant improvements:

**Just running the functional options**

```
❯ benchstat -col /scenario /tmp/with_config.txt
goos: darwin
goarch: arm64
pkg: github.com/DataDog/dd-trace-go/v2/ddtrace/tracer
          │     none     │         WithStartSpanConfig         │
          │    sec/op    │   sec/op     vs base                │
Config-10   169.65n ± 1%   70.17n ± 5%  -58.64% (p=0.000 n=10)

          │    none    │        WithStartSpanConfig         │
          │    B/op    │    B/op     vs base                │
Config-10   448.0 ± 0%   160.0 ± 0%  -64.29% (p=0.000 n=10)

          │    none    │        WithStartSpanConfig         │
          │ allocs/op  │ allocs/op   vs base                │
Config-10   5.000 ± 0%   2.000 ± 0%  -60.00% (p=0.000 n=10)
``` 

**Creating spans (more realistic)**

```
❯ benchstat -col /scenario /tmp/start_span.txt
goos: darwin
goarch: arm64
pkg: github.com/DataDog/dd-trace-go/v2/ddtrace/tracer
                   │    none     │        WithStartSpanConfig         │
                   │   sec/op    │   sec/op     vs base               │
StartSpanConfig-10   1.387µ ± 1%   1.327µ ± 2%  -4.36% (p=0.000 n=10)

                   │     none     │         WithStartSpanConfig          │
                   │     B/op     │     B/op      vs base                │
StartSpanConfig-10   2.390Ki ± 0%   2.108Ki ± 0%  -11.77% (p=0.000 n=10)

                   │    none    │        WithStartSpanConfig         │
                   │ allocs/op  │ allocs/op   vs base                │
StartSpanConfig-10   25.00 ± 0%   22.00 ± 0%  -12.00% (p=0.000 n=10)
``` 

### Reviewer's Checklist

- [x] Changed code has unit tests for its functionality at or near 100% coverage.
- [x] There is a benchmark for any new code, or changes to existing code.
- [ ] If this interacts with the agent in a new way, a system test has been added.

For Datadog employees:

- [ ] If this PR touches code that handles credentials of any kind, such as Datadog API keys, I've requested a review from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.

Unsure? Have a question? Request a review!